+++ src/api/startup.js
@@ -83,8 +83,9 @@ export default function (
     computedCompare = true,
     watchCompare = true,
     watchImmediate = false,
-    alwaysGiveState = true,
     reComputed = true,
+    computedRetKeyDep = true,
+    watchRetKeyDep = true,
   } = {}) {
   try {
     throw new Error();
@@ -99,13 +100,14 @@ export default function (
       ccContext.reComputed = reComputed;
       ccContext.errorHandler = errorHandler;
       const rv = ccContext.runtimeVar;
-      rv.alwaysGiveState = alwaysGiveState;
       rv.isStrict = isStrict;
       rv.isDebug = isDebug;
       rv.computedCompare = computedCompare;
       rv.watchCompare = watchCompare;
       rv.watchImmediate = watchImmediate;
       rv.bindCtxToMethod = bindCtxToMethod;
+      rv.computedRetKeyDep = computedRetKeyDep;
+      rv.watchRetKeyDep = watchRetKeyDep;
 
       if (!ccContext.refs[CC_DISPATCHER]) {
         const Dispatcher = createDispatcher();