+++ src/cc-context/index.js
@@ -2,7 +2,7 @@ import moduleName_stateKeys_ from './statekeys-map';
 import computed from './computed-map';
 import watch from './watch-map';
 import runtimeVar from './runtime-var';
-import waKey_uKeyMap_ from './wakey-ukey-map';
+import { waKey_uKeyMap_, waKey_staticUKeyMap_ } from './wakey-ukey-map';
 import { MODULE_GLOBAL, MODULE_CC, MODULE_DEFAULT, MODULE_VOID, CATE_MODULE, CC_DISPATCHER } from '../support/constant';
 import * as util from '../support/util';
 import pickDepFns from '../core/base/pick-dep-fns';
@@ -217,15 +217,13 @@ const ccContext = {
   handlerKey_handler_: {},
   // { 'foo/f1': {ukey1: 1, ukey2:1 } }
   waKey_uKeyMap_,
-  // 由实例effect依赖列表定义的key记录到此处，写入时机是beforeMount时，是一个实例化完成就会固化的依赖
-  // 由于effect的key较少，所以对waKey_effectUKeyMap_不采取一开始映射好全部waKey的形式，而是采用safeGet动态添加map映射
-  waKey_effectUKeyMap_ : {},
+  waKey_staticUKeyMap_,
   refs,
   info: {
     packageLoadTime: Date.now(),
     firstStartupTime: '',
     latestStartupTime: '',
-    version: '2.3.17',
+    version: '2.3.18',
     author: 'fantasticsoul',
     emails: ['624313307@qq.com', 'zhongzhengkai@gmail.com'],
     tag: 'yuna',