+++ src/cc-context/runtime-var.js
@@ -1,7 +1,6 @@
 
 // 后续在逐步迁移其他的
 export default {
-  alwaysGiveState: true,
   // if isStrict is true, every error will be throw out instead of console.error, 
   // but this may crash your app, make sure you have a nice error handling way,
   // like componentDidCatch in react 16.*
@@ -11,4 +10,6 @@ export default {
   watchCompare: true,
   watchImmediate: false,
   bindCtxToMethod: false,
+  computedRetKeyDep: true,
+  watchRetKeyDep: true,
 }