+++ src/core/base/did-mount.js
@@ -1,16 +1,24 @@
-import * as ev from '../event';
 import triggerSetupEffect from './trigger-setup-effect';
 import setRef from '../ref/set-ref';
 import afterRender from '../ref/after-render';
+import { okeys } from '../../support/util';
+import { mapStaticInsM } from '../../cc-context/wakey-ukey-map';
 
 export default function (ref) {
   afterRender(ref);
 
   ref.__$$isMounted = true;
   ref.__$$isUnmounted = false;
-  const { isSingle, ccClassKey, ccKey, ccUniqueKey, __$$onEvents } = ref.ctx;
+  const { isSingle, ccClassKey, ccKey, ccUniqueKey, __$$onEvents, __$$staticWaKeys } = ref.ctx;
+
   setRef(ref, isSingle, ccClassKey, ccKey, ccUniqueKey);
 
+  const __$$staticWaKeyList = okeys(__$$staticWaKeys);
+  // 用于辅助清理依赖映射
+  ref.ctx.__$$staticWaKeyList = __$$staticWaKeyList;
+  // 记录静态依赖
+  __$$staticWaKeyList.forEach(modStateKey => mapStaticInsM(modStateKey, ccUniqueKey));
+  
   // 这些事件是组件还未挂载时，就派发过来的，延迟到此刻执行，同时清空
   if (__$$onEvents.length > 0) {
     __$$onEvents.forEach(({ fn, args }) => fn(...args));
@@ -18,4 +26,5 @@ export default function (ref) {
   }
 
   triggerSetupEffect(ref, true);
+
 }