+++ src/core/base/trigger-computed-and-watch.js
@@ -1,5 +1,5 @@
 import ccContext from '../../cc-context';
-import { okeys, makeCallInfo } from '../../support/util';
+import { makeCallInfo } from '../../support/util';
 import computeValueForRef from '../computed/compute-value-for-ref';
 import watchKeyForRef from '../watch/watch-key-for-ref';
 
@@ -8,27 +8,22 @@ const { store: { getState } } = ccContext;
 /** 由首次render触发 */
 export default function (ref) {
   const ctx = ref.ctx;
-  const { hasComputedFn, hasWatchFn, connect, module: refModule } = ctx;
+
+  // 取原始对象，防止computeValueForRef里用Object.assign触发依赖收集
+  // 首次挂载组件时，prevState是原始对象，state可能是代理过的对象
+  const { hasComputedFn, hasWatchFn, connectedModules, module: refModule, state: refState } = ctx;
 
   const callInfo = makeCallInfo(refModule);
-  const connectedModules = okeys(connect);
-  const refState = ctx.state;
-  if (hasComputedFn) {
-    computeValueForRef(ref, refModule, refState, refState, callInfo, true, true);
+
+  const cuOrWatch = (op) => {
+    op(ref, refModule, refState, refState, callInfo, true, true);
     connectedModules.forEach(m => {
       const mState = getState(m);
       const tmpCallInfo = makeCallInfo(m);
-      computeValueForRef(ref, m, mState, mState, tmpCallInfo, true, true);
+      op(ref, m, mState, mState, tmpCallInfo, true, true);
     });
   }
 
-  if (hasWatchFn) {
-    watchKeyForRef(ref, refModule, refState, refState, callInfo, true, true);
-    connectedModules.forEach(m => {
-      const mState = getState(m);
-      const tmpCallInfo = makeCallInfo(m);
-      watchKeyForRef(ref, m, mState, mState, tmpCallInfo, true, true);
-    });
-  }
-
+  if (hasComputedFn) cuOrWatch(computeValueForRef);
+  if (hasWatchFn) cuOrWatch(watchKeyForRef);
 }