+++ src/core/ref/after-render.js
@@ -1,9 +1,7 @@
 import { END } from '../../support/priv-constant';
 import { okeys } from '../../support/util';
 import * as cache from './_cache';
-import ccContext from '../../cc-context';
-
-const { waKey_uKeyMap_ } = ccContext;
+import { delIns } from '../../cc-context/wakey-ukey-map';
 
 // before render
 // cur: {} compare: {a:2, b:2, c:2} compareCount=3 nextCompare:{}
@@ -27,9 +25,7 @@ function delDep(compareWaKeys, compareWaKeyCount, module, ccUniqueKey ){
   waKeys.forEach(waKey=>{// no module prefix
     if(compareWaKeys[waKey] === 2 ){//这个key在这轮渲染结束后没有命中，说明视图不再对它有依赖
       shouldLetCacheExpire = true;
-      const prefixedWaKey = `${module}/${waKey}`;
-      // waKey_uKeyMap_[prefixedWaKey][ccUniqueKey] = 2;//处于非依赖状态
-      delete waKey_uKeyMap_[prefixedWaKey][ccUniqueKey];
+      delIns(module, waKey, ccUniqueKey);
     }
   });
   if(waKeys.length > compareWaKeyCount){//大于最初记录的key数量，有新增