+++ src/index.d.ts
@@ -823,9 +823,8 @@ interface RunOptions {
   watchCompare?: boolean;// default is true
   watchImmediate?: boolean;// default is false
   reComputed?: boolean;// default is true
-  alwaysGiveState?: boolean;// default is true
-  clearShadowRef?: boolean;// default is true, but this will not effect when in production mode
-  shawRefExpireTime?: number;// default is 10000
+  computedRetKeyDep?: boolean;
+  watchRetKeyDep?: boolean;
 }
 
 export interface IActionCtxBase {